ref(core): Rename `Hub` to `AsyncContextStack` & remove unneeded methods (#11630)
Should save some bundle size...! I also moved stuff around a bit so we
don't have so much somewhat unrelated stuff packed into the former
`hub.ts` file - needed some work to avoid circular dependencies, but I
think it's fine now.

I left the stuff we actually need in, and renamed it for clarity so this
is not confusing anymore.

closes #11482
